requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
